test(fixtures): modularize git repo file into sub-modules
