+++ tests/fixtures/__init__.py
@@ -1,4 +1,5 @@
 from tests.fixtures.commit_parsers import *
 from tests.fixtures.example_project import *
 from tests.fixtures.git_repo import *
+from tests.fixtures.repos import *
 from tests.fixtures.scipy import *
