+++ tests/fixtures/repos/__init__.py
@@ -1,0 +1,3 @@
+from tests.fixtures.repos.git_flow import *
+from tests.fixtures.repos.github_flow import *
+from tests.fixtures.repos.trunk_based_dev import *
