+++ tests/fixtures/repos/git_flow/__init__.py
@@ -1,0 +1,2 @@
+from tests.fixtures.repos.git_flow.repo_w_2_release_channels import *
+from tests.fixtures.repos.git_flow.repo_w_3_release_channels import *
