+++ tests/fixtures/repos/github_flow/__init__.py
@@ -1,0 +1,1 @@
+from tests.fixtures.repos.github_flow.repo_w_release_channels import *
