+++ tests/fixtures/repos/github_flow/repo_w_release_channels.py
@@ -1,0 +1,239 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+import pytest
+
+from tests.const import COMMIT_MESSAGE
+from tests.util import add_text_to_file
+
+if TYPE_CHECKING:
+    from git import Repo
+
+    from tests.fixtures.example_project import UpdatePyprojectTomlFn, UseParserFn
+    from tests.fixtures.git_repo import RepoInitFn
+
+
+@pytest.fixture
+def repo_with_main_and_feature_branches_angular_commits(
+    git_repo_factory: RepoInitFn,
+    use_angular_parser: UseParserFn,
+    update_pyproject_toml: UpdatePyprojectTomlFn,
+    file_in_repo: str,
+    change_to_ex_proj_dir: None,
+) -> Repo:
+    git_repo = git_repo_factory()
+    use_angular_parser()
+    update_pyproject_toml(
+        "tool.semantic_release.branches.beta-testing",
+        {"match": "beta.*", "prerelease": True, "prerelease_token": "beta"},
+    )
+
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m="Initial commit")
+
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m=COMMIT_MESSAGE.format(version="0.1.0"))
+    git_repo.git.tag("v0.1.0", m="v0.1.0")
+
+    # Do a prerelease
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m="fix: add some more text")
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m=COMMIT_MESSAGE.format(version="0.1.1-rc.1"))
+    git_repo.git.tag("v0.1.1-rc.1", m="v0.1.1-rc.1")
+
+    # Do a prerelease
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m="feat: add some more text")
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m=COMMIT_MESSAGE.format(version="0.2.0-rc.1"))
+    git_repo.git.tag("v0.2.0-rc.1", m="v0.2.0-rc.1")
+
+    # Do a full release
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m="feat: add some more text")
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m=COMMIT_MESSAGE.format(version="0.2.0"))
+    git_repo.git.tag("v0.2.0", m="v0.2.0")
+
+    assert git_repo.commit("v0.2.0").hexsha == git_repo.head.commit.hexsha
+
+    git_repo.create_head("beta_testing")
+    git_repo.heads.beta_testing.checkout()
+
+    # Do a prerelease on the branch
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m="feat: (feature) add some more text")
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m=COMMIT_MESSAGE.format(version="0.3.0-beta.1"))
+    git_repo.git.tag("v0.3.0-beta.1", m="v0.3.0-beta.1")
+
+    assert git_repo.commit("v0.3.0-beta.1").hexsha == git_repo.head.commit.hexsha
+    assert git_repo.active_branch.name == "beta_testing"
+    return git_repo
+
+
+@pytest.fixture
+def repo_with_main_and_feature_branches_emoji_commits(
+    git_repo_factory: RepoInitFn,
+    use_emoji_parser: UseParserFn,
+    file_in_repo: str,
+    change_to_ex_proj_dir: None,
+) -> Repo:
+    git_repo = git_repo_factory()
+    use_emoji_parser()
+
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m="Initial commit")
+
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m=COMMIT_MESSAGE.format(version="0.1.0"))
+    git_repo.git.tag("v0.1.0", m="v0.1.0")
+
+    # Do a prerelease
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m=":bug: add some more text")
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m=COMMIT_MESSAGE.format(version="0.1.1-rc.1"))
+    git_repo.git.tag("v0.1.1-rc.1", m="v0.1.1-rc.1")
+
+    # Do a prerelease
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m=":sparkles: add some more text")
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m=COMMIT_MESSAGE.format(version="0.2.0-rc.1"))
+    git_repo.git.tag("v0.2.0-rc.1", m="v0.2.0-rc.1")
+
+    # Do a full release
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m=":sparkles: add some more text")
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m=COMMIT_MESSAGE.format(version="0.2.0"))
+    git_repo.git.tag("v0.2.0", m="v0.2.0")
+
+    assert git_repo.commit("v0.2.0").hexsha == git_repo.head.commit.hexsha
+
+    git_repo.create_head("beta_testing")
+    git_repo.heads.beta_testing.checkout()
+
+    # Do a prerelease on the branch
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m=":sparkles: (feature) add some more text")
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m=COMMIT_MESSAGE.format(version="0.3.0-beta.1"))
+    git_repo.git.tag("v0.3.0-beta.1", m="v0.3.0-beta.1")
+
+    assert git_repo.commit("v0.3.0-beta.1").hexsha == git_repo.head.commit.hexsha
+    assert git_repo.active_branch.name == "beta_testing"
+    return git_repo
+
+
+@pytest.fixture
+def repo_with_main_and_feature_branches_scipy_commits(
+    git_repo_factory: RepoInitFn,
+    use_scipy_parser: UseParserFn,
+    file_in_repo: str,
+    change_to_ex_proj_dir: None,
+) -> Repo:
+    git_repo = git_repo_factory()
+    use_scipy_parser()
+
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m="Initial commit")
+
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m=COMMIT_MESSAGE.format(version="0.1.0"))
+    git_repo.git.tag("v0.1.0", m="v0.1.0")
+
+    # Do a prerelease
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m="MAINT: add some more text")
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m=COMMIT_MESSAGE.format(version="0.1.1-rc.1"))
+    git_repo.git.tag("v0.1.1-rc.1", m="v0.1.1-rc.1")
+
+    # Do a prerelease
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m="ENH: add some more text")
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m=COMMIT_MESSAGE.format(version="0.2.0-rc.1"))
+    git_repo.git.tag("v0.2.0-rc.1", m="v0.2.0-rc.1")
+
+    # Do a full release
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m="ENH: add some more text")
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m=COMMIT_MESSAGE.format(version="0.2.0"))
+    git_repo.git.tag("v0.2.0", m="v0.2.0")
+
+    assert git_repo.commit("v0.2.0").hexsha == git_repo.head.commit.hexsha
+
+    git_repo.create_head("beta_testing")
+    git_repo.heads.beta_testing.checkout()
+
+    # Do a prerelease on the branch
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m="ENH: (feature) add some more text")
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m=COMMIT_MESSAGE.format(version="0.3.0-beta.1"))
+    git_repo.git.tag("v0.3.0-beta.1", m="v0.3.0-beta.1")
+
+    assert git_repo.commit("v0.3.0-beta.1").hexsha == git_repo.head.commit.hexsha
+    assert git_repo.active_branch.name == "beta_testing"
+    return git_repo
+
+
+@pytest.fixture
+def repo_with_main_and_feature_branches_tag_commits(
+    git_repo_factory: RepoInitFn,
+    use_tag_parser: UseParserFn,
+    file_in_repo: str,
+    change_to_ex_proj_dir: None,
+) -> Repo:
+    git_repo = git_repo_factory()
+    use_tag_parser()
+
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m="Initial commit")
+
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m=COMMIT_MESSAGE.format(version="0.1.0"))
+    git_repo.git.tag("v0.1.0", m="v0.1.0")
+
+    # Do a prerelease
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m=":nut_and_bolt: add some more text")
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m=COMMIT_MESSAGE.format(version="0.1.1-rc.1"))
+    git_repo.git.tag("v0.1.1-rc.1", m="v0.1.1-rc.1")
+
+    # Do a prerelease
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m=":sparkles: add some more text")
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m=COMMIT_MESSAGE.format(version="0.2.0-rc.1"))
+    git_repo.git.tag("v0.2.0-rc.1", m="v0.2.0-rc.1")
+
+    # Do a full release
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m=":sparkles: add some more text")
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m=COMMIT_MESSAGE.format(version="0.2.0"))
+    git_repo.git.tag("v0.2.0", m="v0.2.0")
+
+    assert git_repo.commit("v0.2.0").hexsha == git_repo.head.commit.hexsha
+
+    git_repo.create_head("beta_testing")
+    git_repo.heads.beta_testing.checkout()
+
+    # Do a prerelease on the branch
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m=":sparkles: (feature) add some more text")
+    add_text_to_file(git_repo, file_in_repo)
+    git_repo.git.commit(m=COMMIT_MESSAGE.format(version="0.3.0-beta.1"))
+    git_repo.git.tag("v0.3.0-beta.1", m="v0.3.0-beta.1")
+
+    assert git_repo.commit("v0.3.0-beta.1").hexsha == git_repo.head.commit.hexsha
+    assert git_repo.active_branch.name == "beta_testing"
+    return git_repo
