+++ tests/fixtures/repos/trunk_based_dev/__init__.py
@@ -1,0 +1,3 @@
+from tests.fixtures.repos.trunk_based_dev.repo_w_no_tags import *
+from tests.fixtures.repos.trunk_based_dev.repo_w_prereleases import *
+from tests.fixtures.repos.trunk_based_dev.repo_w_tags import *
